+++ sim/src/insn_impl/csrrwi.rs
@@ -6,7 +6,7 @@ use crate::{
 pub fn csrrwi(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     let rd = insn.rd();
     let csr = insn.csr();
-    let zimm = insn.zimm() as u64;
+    let zimm = insn.zimm();
 
     let csr_value = cpu.csrs.load(csr)?;
     cpu.csrs.store(csr, zimm)?;
