+++ sim/src/insn_impl/divu.rs
@@ -14,7 +14,7 @@ pub fn divu(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     if rs2_value == 0 {
         cpu.store(rd, u64::MAX);
     } else {
-        cpu.store(rd, (rs1_value / rs2_value) as u64);
+        cpu.store(rd, rs1_value / rs2_value);
     }
 
     Ok(cpu.pc + 4)
