+++ sim/src/insn_impl/fsd.rs
@@ -8,7 +8,7 @@ pub fn fsd(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     let rs2 = insn.rs2();
     let imm12lo = insn.imm12lo();
     
-    let offset = Insn::sign_extend((imm12hi << 5 | imm12lo) as u64, 12);
+    let offset = Insn::sign_extend((imm12hi << 5) | imm12lo, 12);
 
     insn_raw::fsd_raw::fsd_raw(cpu, bus, rs1, rs2, offset as u64)
 }
