+++ sim/src/insn_impl/insn_raw/c_li_raw.rs
@@ -2,6 +2,6 @@ use crate::{cpu::{self, Cpu, Insn}, bus::Bus};
 
 #[inline(always)]
 pub fn c_li_raw(cpu: &mut Cpu, rd_n0: u64, imm_c_li: u64) -> cpu::Result<u64> {
-    cpu.store(rd_n0, imm_c_li as u64);
+    cpu.store(rd_n0, imm_c_li);
     Ok(cpu.pc + 2)
 }
