+++ sim/src/insn_impl/insn_raw/divu_raw.rs
@@ -8,7 +8,7 @@ pub fn divu_raw(cpu: &mut Cpu, rd: u64, rs1: u64, rs2: u64) -> cpu::Result<u64> 
     if rs2_value == 0 {
         cpu.store(rd, u64::MAX);
     } else {
-        cpu.store(rd, (rs1_value / rs2_value) as u64);
+        cpu.store(rd, rs1_value / rs2_value);
     }
 
     Ok(cpu.pc + 4)
