+++ sim/src/insn_impl/insn_raw/fcvt_lu_d_raw.rs
@@ -23,6 +23,6 @@ pub fn fcvt_lu_d_raw(cpu: &mut Cpu, rd: u64, rs1: u64, rm: u64) -> cpu::Result<u
     }
 }
     cpu.set_fflags();
-    cpu.store(rd, result as u64);
+    cpu.store(rd, result);
     Ok(cpu.pc + 4)
 }
