+++ sim/src/insn_impl/insn_raw/fld_raw.rs
@@ -4,8 +4,8 @@ use crate::{bus::{Bus, Device}, cpu::{self, Cpu, Insn}};
 
 #[inline(always)]
 pub fn fld_raw(cpu: &mut Cpu, bus: &mut Bus, rd: u64, rs1: u64, imm_i: u64) -> cpu::Result<u64> {
-    let imm = Insn::sign_extend(imm_i as u64, 12);
-    let address = (cpu.load(rs1) as u64).wrapping_add(imm as u64);
+    let imm = Insn::sign_extend(imm_i, 12);
+    let address = cpu.load(rs1).wrapping_add(imm as u64);
 
     let mut raw = [0; size_of::<u64>()];
     bus.read(address, &mut raw)?;
