+++ sim/src/insn_impl/mulw.rs
@@ -9,7 +9,7 @@ pub fn mulw(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     let rs2 = insn.rs2();
 
     let result = Insn::sign_extend(
-        (cpu.load(rs1) as u32).wrapping_mul(cpu.load(rs2) as u32) as u32 as u64,
+        (cpu.load(rs1) as u32).wrapping_mul(cpu.load(rs2) as u32) as u64,
         32,
     );
     cpu.store(rd, result as u64);
